+++ demo/src/main/java/com/example/demo/IdeaPackage/IdeaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IdeaRepository extends JpaRepository<Idea,Long> {
+public interface IdeaRepository extends JpaRepository<Idea, Integer> {
 }
